requirements.txt
openai>=1.55.0
python-dotenv>=1.0.1
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.26.0
duckduckgo-search>=6.3.3
rich>=13.7.1
